refactor: Renamed daily habit tracker file
